Oddělení serverového vlákna do samostatné třídy.
