+++ drd_server/src/main/java/cz/stechy/drd/Server.java
@@ -3,22 +3,13 @@ package cz.stechy.drd;
 import com.google.auth.oauth2.GoogleCredentials;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseOptions;
-import cz.stechy.drd.net.message.HelloMessage;
-import cz.stechy.drd.net.message.MessageSource;
 import java.io.FileInputStream;
 import java.io.IOException;
-import java.net.ServerSocket;
-import java.net.Socket;
-import java.net.SocketTimeoutException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Scanner;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class Server extends Thread {
+public class Server {
 
     @SuppressWarnings("unused")
     private static final Logger LOGGER = LoggerFactory.getLogger(Server.class);
@@ -33,39 +24,21 @@ public class Server extends Thread {
 
     // Nastavení serveru
     private final CmdParser settings;
-    private final int maxClients;
+    private final ServerThread serverThread;
 
-    ExecutorService pool;
-    private boolean running = false;
-    private final List<Client> clients = new ArrayList<>();
-    private final ClientDispatcher clientDispatcher;
-    private final WriterThread writerThread;
-
-    public static void main(String[] args) throws InterruptedException {
+    public static void main(String[] args) {
         final Server server = new Server(new CmdParser().parse(args));
-        server.startListening();
-        // Výchozí vlákno
-        while(true) {
-            final String line = scanner.nextLine();
-            if (line.equals("end")) {
-                LOGGER.info("Spouštím ukončovací sekvenci.");
-                server.shutdown();
-                server.join();
-                LOGGER.info("Server byl ukončen.");
-                break;
-            }
-        }
+        server.run();
     }
 
     private Server(final CmdParser parser) {
-        super("ServerThread");
         LOGGER.info("Spouštím server.");
         this.settings = parser;
-        this.maxClients = settings.getInteger(CmdParser.CLIENTS_COUNT, DEFAULT_MAX_CLIENTS);
+        final int port = settings.getInteger(CmdParser.PORT, SERVER_PORT);
+        final int maxClients = settings.getInteger(CmdParser.CLIENTS_COUNT, DEFAULT_MAX_CLIENTS);
         final int waitingQueueSize = settings.getInteger(CmdParser.MAX_WAITING_QUEUE, DEFAULT_WAITING_QUEUE_SIZE);
+        this.serverThread = new ServerThread(port, maxClients, waitingQueueSize);
 
-        this.clientDispatcher = new ClientDispatcher(waitingQueueSize);
-        this.writerThread = new WriterThread();
         init();
     }
 
@@ -101,84 +74,21 @@ public class Server extends Thread {
         LOGGER.info("Inicializace dokončena.");
     }
 
-    /**
-     * Spustí naslouchání serveru
-     */
-    private void startListening() {
-        LOGGER.info("Inicializuji klientská vlákna.");
-        pool = Executors.newFixedThreadPool(maxClients);
-        running = true;
-        clientDispatcher.start();
-        writerThread.start();
-        super.start();
-    }
-
-    public void shutdown() {
-        running = false;
-    }
-
-    private synchronized void insertClientToListOrQueue(Client client) {
-        if (clients.size() < maxClients) {
-            LOGGER.info("Přidávám klienta do kolekce klientů a spouštím komunikaci.");
-            clients.add(client);
-            client.setConnectionClosedListener(() -> {
-                LOGGER.info("Odebírám klienta ze seznamu na serveru.");
-                clients.remove(client);
-                LOGGER.info("Počet připojených klientů: " + clients.size());
-                if (clientDispatcher.hasClientInQueue()) {
-                    LOGGER.info("V čekací listině se našel klient, který by rád komunikoval.");
-                    this.insertClientToListOrQueue(clientDispatcher.getClientFromQueue());
-                }
-            });
-            pool.submit(client);
-            client.sendMessage(new HelloMessage(MessageSource.SERVER));
-        } else {
-            if (clientDispatcher.addClientToQueue(client)) {
-                LOGGER.info("Přidávám klienta na čekací listinu.");
-            } else {
-                LOGGER.warn("Odpojuji klienta od serveru. Je připojeno příliš mnoho uživatelů.");
-                client.close();
+    private void run() {
+        serverThread.start();
+        while(true) {
+            final String line = scanner.nextLine();
+            if (line.equals("end")) {
+                LOGGER.info("Spouštím ukončovací sekvenci.");
+                break;
             }
         }
-    }
 
-    @Override
-    public void run() {
-        try (ServerSocket serverSocket = new ServerSocket(SERVER_PORT)) {
-            serverSocket.setSoTimeout(5000);
-            LOGGER.info(String.format("Server naslouchá na portu: %d.", serverSocket.getLocalPort()));
-            // Nové vlákno serveru
-            while (running) {
-                try {
-                    final Socket socket = serverSocket.accept();
-                    LOGGER.info("Server přijal nové spojení.");
+        serverThread.shutdown();
+        try {
+            serverThread.join();
+        } catch (InterruptedException e) {}
 
-                    final Client client = new Client(socket, writerThread);
-                    this.insertClientToListOrQueue(client);
-                } catch (SocketTimeoutException e) {}
-            }
-
-            LOGGER.info("Ukončuji server.");
-            LOGGER.info("Odpojuji připojené klienty.");
-            for (Client client : clients) {
-                client.disconnect();
-            }
-            LOGGER.info("Ukončuji činnost thread poolu.");
-            pool.shutdown();
-            LOGGER.info("Ukončuji client dispatcher.");
-            clientDispatcher.shutdown();
-            try {
-                clientDispatcher.join();
-            } catch (InterruptedException e) {}
-            LOGGER.info("Ukončuji writer thread");
-            writerThread.shutdown();
-            try {
-                writerThread.join();
-            } catch (InterruptedException e) {}
-            LOGGER.info("Naslouchací vlákno bylo úspěšně ukončeno.");
-
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
+        LOGGER.info("Server byl ukončen.");
     }
 }
